Move storybook utils back into .stories file
The pattern of moving them out made them become part of the build
and forced dev deps to need to be prod deps.
